Resolve installed module
The `npm install` command doesn't provide a return that lets you determine which module was installed.  When providing a single argument, multiple modules may be installed if there are peer dependencies.  And because the provided argument can be a URL or file path, it won't necessarily contain the name of the module to install (that is determined from the package.json).  So we have to look through all of the installed modules (in the case of more than one) and figure out which one is not a peer dependency of the others.